requirements.txt
django==3.2.25
djongo==1.3.6
djangorestframework==3.15.1
Pyfhel==3.4.2
numpy==1.21.0
